requirements.txt
fastapi
uvicorn
opengradient
httpx
pydantic
python-dotenv
tavily-python
